requirements.txt
Django==4.2.3
djangorestframework==3.14.0
python-dotenv==1.0.0
drf_yasg==1.21.6
coverage==7.2.7
